fix: explicit default for attrs.field
